+++ packages/orq-rc/src/models/operations/updateprompt.ts
@@ -1330,12 +1330,8 @@ export type UpdatePromptPromptsResponseMetadata = {
  */
 export type UpdatePromptVersions = {
   id: string;
-  createdById: string;
-  /**
-   * The prompt’s name, meant to be displayable in the UI.
-   */
-  displayName: string;
-  updatedById: string;
+  createdById?: string | undefined;
+  updatedById?: string | undefined;
   /**
    * The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose
    */
@@ -1355,12 +1351,12 @@ export type UpdatePromptResponseBody = {
   id: string;
   owner: string | UpdatePromptOwner2;
   domainId: string;
-  createdById: string;
+  createdById?: string | undefined;
   /**
    * The prompt’s name, meant to be displayable in the UI.
    */
   displayName: string;
-  updatedById: string;
+  updatedById?: string | undefined;
   /**
    * The prompt’s description, meant to be displayable in the UI. Use this field to optionally store a long form explanation of the prompt for your own purpose
    */
@@ -5912,9 +5908,8 @@ export const UpdatePromptVersions$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   _id: z.string(),
-  created_by_id: z.string(),
-  display_name: z.string(),
-  updated_by_id: z.string(),
+  created_by_id: z.string().optional(),
+  updated_by_id: z.string().optional(),
   description: z.nullable(z.string()).optional(),
   prompt_config: z.lazy(() =>
     UpdatePromptPromptsResponsePromptConfig$inboundSchema
@@ -5925,7 +5920,6 @@ export const UpdatePromptVersions$inboundSchema: z.ZodType<
   return remap$(v, {
     "_id": "id",
     "created_by_id": "createdById",
-    "display_name": "displayName",
     "updated_by_id": "updatedById",
     "prompt_config": "promptConfig",
   });
@@ -5934,9 +5928,8 @@ export const UpdatePromptVersions$inboundSchema: z.ZodType<
 /** @internal */
 export type UpdatePromptVersions$Outbound = {
   _id: string;
-  created_by_id: string;
-  display_name: string;
-  updated_by_id: string;
+  created_by_id?: string | undefined;
+  updated_by_id?: string | undefined;
   description?: string | null | undefined;
   prompt_config: UpdatePromptPromptsResponsePromptConfig$Outbound;
   metadata: UpdatePromptPromptsResponseMetadata$Outbound;
@@ -5950,9 +5943,8 @@ export const UpdatePromptVersions$outboundSchema: z.ZodType<
   UpdatePromptVersions
 > = z.object({
   id: z.string(),
-  createdById: z.string(),
-  displayName: z.string(),
-  updatedById: z.string(),
+  createdById: z.string().optional(),
+  updatedById: z.string().optional(),
   description: z.nullable(z.string()).optional(),
   promptConfig: z.lazy(() =>
     UpdatePromptPromptsResponsePromptConfig$outboundSchema
@@ -5963,7 +5955,6 @@ export const UpdatePromptVersions$outboundSchema: z.ZodType<
   return remap$(v, {
     id: "_id",
     createdById: "created_by_id",
-    displayName: "display_name",
     updatedById: "updated_by_id",
     promptConfig: "prompt_config",
   });
@@ -6009,16 +6000,16 @@ export const UpdatePromptResponseBody$inboundSchema: z.ZodType<
   _id: z.string(),
   owner: z.union([z.string(), UpdatePromptOwner2$inboundSchema]),
   domain_id: z.string(),
-  created_by_id: z.string(),
+  created_by_id: z.string().optional(),
   display_name: z.string(),
-  updated_by_id: z.string(),
+  updated_by_id: z.string().optional(),
   description: z.nullable(z.string()).optional(),
   prompt_config: z.lazy(() => UpdatePromptPromptsPromptConfig$inboundSchema),
   metadata: z.lazy(() => UpdatePromptPromptsMetadata$inboundSchema),
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-02-10T14:13:16.580Z",
+    "2025-02-10T14:20:21.534Z",
   ).transform(v => new Date(v)),
   type: UpdatePromptPromptsType$inboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptVersions$inboundSchema)),
@@ -6038,9 +6029,9 @@ export type UpdatePromptResponseBody$Outbound = {
   _id: string;
   owner: string | string;
   domain_id: string;
-  created_by_id: string;
+  created_by_id?: string | undefined;
   display_name: string;
-  updated_by_id: string;
+  updated_by_id?: string | undefined;
   description?: string | null | undefined;
   prompt_config: UpdatePromptPromptsPromptConfig$Outbound;
   metadata: UpdatePromptPromptsMetadata$Outbound;
@@ -6059,14 +6050,14 @@ export const UpdatePromptResponseBody$outboundSchema: z.ZodType<
   id: z.string(),
   owner: z.union([z.string(), UpdatePromptOwner2$outboundSchema]),
   domainId: z.string(),
-  createdById: z.string(),
+  createdById: z.string().optional(),
   displayName: z.string(),
-  updatedById: z.string(),
+  updatedById: z.string().optional(),
   description: z.nullable(z.string()).optional(),
   promptConfig: z.lazy(() => UpdatePromptPromptsPromptConfig$outboundSchema),
   metadata: z.lazy(() => UpdatePromptPromptsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-02-10T14:13:16.580Z"))
+  updated: z.date().default(() => new Date("2025-02-10T14:20:21.534Z"))
     .transform(v => v.toISOString()),
   type: UpdatePromptPromptsType$outboundSchema,
   versions: z.array(z.lazy(() => UpdatePromptVersions$outboundSchema)),