+++ packages/orq-rc/RELEASES.md
@@ -309,3 +309,13 @@ Based on:
 - [typescript v3.1.0-rc.51] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.51 - packages/orq-rc
+
+## 2025-02-10 14:21:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.487.1 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.52] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.52 - packages/orq-rc